[MEDI-58] refactor: refactoring notification API. 7/31.
