+++ src/main/java/com/my_medi/api/consultation/service/ConsultationUseCase.java
@@ -4,9 +4,11 @@ import com.my_medi.domain.consultationRequest.entity.ConsultationRequest;
 import com.my_medi.domain.consultationRequest.service.ConsultationRequestCommandService;
 import com.my_medi.domain.consultationRequest.service.ConsultationRequestQueryService;
 import com.my_medi.domain.expert.entity.Expert;
-import com.my_medi.domain.expertNotification.service.ExpertNotificationCommandService;
+import com.my_medi.domain.notification.entity.NotificationMessage;
+import com.my_medi.domain.notification.entity.NotificationType;
+import com.my_medi.domain.notification.service.ExpertNotificationCommandService;
 import com.my_medi.domain.user.entity.User;
-import com.my_medi.domain.userNotification.service.UserNotificationCommandService;
+import com.my_medi.domain.notification.service.UserNotificationCommandService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -22,14 +24,19 @@ public class ConsultationUseCase {
         consultationRequestCommandService.approveConsultation(consultationId, expert);
         ConsultationRequest request = consultationRequestQueryService.getRequestById(consultationId);
 
-        userNotificationCommandService.
-                sendConsultationRequestApproveNotificationToUser(request.getUser().getId(), consultationId);
+        String notificationComment = NotificationMessage
+                .CONSULTATION_APPROVED.format(request.getUser().getName());
+
+        userNotificationCommandService.sendNotificationToUser(request.getUser().getId(),
+                consultationId, notificationComment, NotificationType.CONSULTATION_RESPONSE);
     }
 
     public Long sendConsultationRequestNotificationToExpert(User user, Long expertId, String comment) {
         Long requestId = consultationRequestCommandService.requestConsultationToExpert(user, expertId, comment);
 
-        expertNotificationCommandService.sendNotificationToExpert(expertId, requestId);
+        String notificationComment = NotificationMessage.CONSULTATION_REQUESTED.format(user.getName());
+
+        expertNotificationCommandService.sendNotificationToExpert(expertId, requestId, notificationComment);
 
         return requestId;
     }
