+++ src/main/java/com/my_medi/api/expertNotification/dto/ExpertNotificationResponseDto.java
@@ -1,18 +1,18 @@
-package com.my_medi.api.userNotification.dto;
+package com.my_medi.api.expertNotification.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-public class UserNotificationResponseDto {
+public class ExpertNotificationResponseDto {
     @Data
     @Builder
     @NoArgsConstructor
     @AllArgsConstructor
-    public static class UserNotificationDto {
+    public static class ExpertNotificationDto {
         private Long id;
-        private Long userId;
+        private Long expertId;
         private String notificationContent;
         private Long sourceId;
         private Boolean isRead;
