+++ src/main/java/com/my_medi/api/userNotification/controller/UserNotificationApiController.java
@@ -5,8 +5,8 @@ import com.my_medi.api.common.dto.ApiResponseDto;
 import com.my_medi.api.userNotification.mapper.UserNotificationConverter;
 import com.my_medi.common.annotation.AuthUser;
 import com.my_medi.domain.user.entity.User;
-import com.my_medi.domain.userNotification.entity.UserNotification;
-import com.my_medi.domain.userNotification.service.UserNotificationQueryService;
+import com.my_medi.domain.notification.entity.UserNotification;
+import com.my_medi.domain.notification.service.UserNotificationQueryService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
