+++ src/main/java/com/my_medi/domain/notification/entity/NotificationType.java
@@ -1,0 +1,5 @@
+package com.my_medi.domain.notification.entity;
+
+public enum NotificationType {
+    CONSULTATION_RESPONSE, SCHEDULE
+}
