+++ src/main/java/com/my_medi/domain/notification/exception/UserNotificationErrorStatus.java
@@ -1,0 +1,48 @@
+package com.my_medi.domain.notification.exception;
+
+import com.my_medi.common.annotation.ExplainError;
+import com.my_medi.common.exception.BaseErrorCode;
+import com.my_medi.common.exception.Reason;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import org.springframework.http.HttpStatus;
+
+import java.lang.reflect.Field;
+import java.util.Objects;
+
+@Getter
+@AllArgsConstructor
+public enum UserNotificationErrorStatus implements BaseErrorCode {
+
+    USER_NOTIFICATION_NOT_FOUND(HttpStatus.BAD_REQUEST, 4250, "user 알림을 찾을 수 없습니다.");
+
+    private final HttpStatus httpStatus;
+    private final Integer code;
+    private final String message;
+
+    @Override
+    public Reason getReason() {
+        return Reason.builder()
+                .message(message)
+                .code(code)
+                .isSuccess(false)
+                .build();
+    }
+
+    @Override
+    public Reason getReasonHttpStatus() {
+        return Reason.builder()
+                .message(message)
+                .code(code)
+                .isSuccess(false)
+                .httpStatus(httpStatus)
+                .build();
+    }
+
+    @Override
+    public String getExplainError() throws NoSuchFieldException {
+        Field field = this.getClass().getField(this.name());
+        ExplainError annotation = field.getAnnotation(ExplainError.class);
+        return Objects.nonNull(annotation) ? annotation.value() : this.getMessage();
+    }
+}
