+++ src/main/java/com/my_medi/domain/notification/repository/ExpertNotificationRepository.java
@@ -1,0 +1,10 @@
+package com.my_medi.domain.notification.repository;
+
+import com.my_medi.domain.notification.entity.ExpertNotification;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface ExpertNotificationRepository extends JpaRepository<ExpertNotification, Long> {
+    List<ExpertNotification> findByExpertId(Long expertId);
+}
