+++ src/main/java/com/my_medi/domain/notification/service/ExpertNotificationCommandService.java
@@ -1,0 +1,7 @@
+package com.my_medi.domain.notification.service;
+
+public interface ExpertNotificationCommandService {
+    void sendNotificationToExpert(Long expertId, Long sourceId, String comment);
+
+    void removeNotification(Long notificationId);
+}
