+++ src/main/java/com/my_medi/domain/notification/service/ExpertNotificationQueryService.java
@@ -1,0 +1,11 @@
+package com.my_medi.domain.notification.service;
+
+import com.my_medi.domain.notification.entity.ExpertNotification;
+import java.util.List;
+
+public interface ExpertNotificationQueryService {
+
+    List<ExpertNotification> getNotificationByExpertId(Long expertId);
+
+    ExpertNotification getNotificationById(Long notificationId);
+}
