--- src/main/java/com/my_medi/domain/userNotification/entity/NotificationType.java
@@ -1,5 +1,0 @@
-package com.my_medi.domain.userNotification.entity;
-
-public enum NotificationType {
-    CONSULTATION_RESPONSE, SCHEDULE
-}
