--- src/main/java/com/my_medi/domain/userNotification/repository/UserNotificationRepository.java
@@ -1,10 +1,0 @@
-package com.my_medi.domain.userNotification.repository;
-
-import com.my_medi.domain.userNotification.entity.UserNotification;
-import org.springframework.data.jpa.repository.JpaRepository;
-import java.util.List;
-
-
-public interface UserNotificationRepository extends JpaRepository<UserNotification, Long> {
-    List<UserNotification> findByUserId(Long userId);
-}
