--- src/main/java/com/my_medi/domain/userNotification/service/UserNotificationQueryService.java
@@ -1,10 +1,0 @@
-package com.my_medi.domain.userNotification.service;
-
-import com.my_medi.domain.userNotification.entity.UserNotification;
-import java.util.List;
-
-public interface UserNotificationQueryService {
-    List<UserNotification> getNotificationByUserId(Long userId);
-
-    UserNotification getNotificationById(Long notificationId);
-}
